Move initialization helpers to ParallelAlgorithms
