+++ src/Evolution/Initialization/Domain.hpp
@@ -21,8 +21,8 @@
 #include "Domain/MinimumGridSpacing.hpp"
 #include "Domain/Tags.hpp"
 #include "Evolution/Initialization/Helpers.hpp"
-#include "Evolution/Initialization/MergeIntoDataBox.hpp"
 #include "Evolution/Initialization/Tags.hpp"
+#include "ParallelAlgorithms/Initialization/MergeIntoDataBox.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TaggedTuple.hpp"
