+++ tests/Unit/Evolution/CMakeLists.txt
@@ -5,7 +5,6 @@ add_subdirectory(Actions)
 add_subdirectory(Conservative)
 add_subdirectory(DiscontinuousGalerkin)
 add_subdirectory(EventsAndTriggers)
-add_subdirectory(Initialization)
 add_subdirectory(Systems)
 add_subdirectory(VariableFixing)
