+++ tests/Unit/ParallelAlgorithms/CMakeLists.txt
@@ -2,3 +2,4 @@
 # See LICENSE.txt for details.
 
 add_subdirectory(Actions)
+add_subdirectory(Initialization)
